Add access to default Variable value onto AggressiveSolverManager Split out IndividualSetup so that the VariableStore can be customized
